Move error-code handling to OS layer
Move error-code handling functions, namely, errno_set, errno_get, and
strerror to OS layer.
