+++ include/jemalloc/internal/util.h
@@ -3,6 +3,7 @@
 
 #include "jemalloc/internal/jemalloc_preamble.h"
 #include "jemalloc/internal/jemalloc_internal_types.h"
+#include "jemalloc/internal/os/error.h"
 
 #define UTIL_INLINE static inline
 
@@ -66,21 +67,13 @@ max_zu(size_t a, size_t b) {
 /* Set error code. */
 UTIL_INLINE void
 set_errno(int errnum) {
-#ifdef _WIN32
-	SetLastError(errnum);
-#else
-	errno = errnum;
-#endif
+	os_errno_set(errnum);
 }
 
 /* Get last error code. */
 UTIL_INLINE int
 get_errno(void) {
-#ifdef _WIN32
-	return GetLastError();
-#else
-	return errno;
-#endif
+	return os_errno_get();
 }
 
 #ifdef _MSC_VER
